Koshin Icons
Not sure whether you wanted the civ icons in colour or alpha, so I've
done both to be on the safe side.

The civ icons are exactly the right size. The leader icons don't because
of the horns (the circle is the right size [give or take a pixel] but
the whole file is a bit bigger).